Adapt to upstream DPF restructuring: pugl path, VST2 macro, sub-pixel events
Three independent build fixes for the DPF bump:

1. Add pugl include path. Upstream now uses pugl as a separate
   submodule at dgl/src/pugl-upstream/. The DPF headers reach for
   <pugl/pugl.h> via the include path; teach our forked
   Makefile.plugins.mk about the new location, matching upstream
   Makefile.plugins.mk:101.

2. Rename _VST -> _VST2 in the vst target dependencies. The %
   pattern rule turns the filename suffix into the
   -DDISTRHO_PLUGIN_TARGET_$* macro; DPF renamed the macro from
   DISTRHO_PLUGIN_TARGET_VST to DISTRHO_PLUGIN_TARGET_VST2 to make
   room for VST3.

3. Widen tabview/xypad coordinate types to Point<double>. DPF's
   mouse/motion event 'pos' fields are now Point<double> for
   sub-pixel precision, so isInTabArea() and updateValueFromPos()
   must accept Point<double> too. Update XYPad::cursor accordingly
   and switch the two std::clamp<int>() calls to <double> so the
   double arg binds without narrowing.

+++ plugin/dpf/common/gui/xypad.hpp
@@ -28,7 +28,7 @@ private:
 
   std::vector<double> defaultValue;
 
-  Point<int> cursor{-1, -1};
+  Point<double> cursor{-1, -1};
   bool isMouseEntered = false;
   bool wasMouseEntered = false;
   bool isMouseLeftDown = false;
@@ -146,10 +146,10 @@ public:
   }
 
 private:
-  void updateValueFromPos(const Point<int> &pos)
+  void updateValueFromPos(const Point<double> &pos)
   {
-    value[0] = std::clamp<int>(pos.getX(), 0, getWidth()) / double(getWidth());
-    value[1] = std::clamp<int>(pos.getY(), 0, getHeight()) / double(getHeight());
+    value[0] = std::clamp<double>(pos.getX(), 0.0, double(getWidth())) / double(getWidth());
+    value[1] = std::clamp<double>(pos.getY(), 0.0, double(getHeight())) / double(getHeight());
     updateValue();
   }
 };
